refactor(middleware): move `AuthRequiredMiddleware` from inside auth module into  middleware package
This is done to put middlewares in the same package and auth middleware is the only file inside auth module
